+++ auth-module/src/main/java/com/inhabas/api/auth/domain/socialAccount/SocialAccountRepository.java
@@ -1,6 +1,6 @@
 package com.inhabas.api.auth.domain.socialAccount;
 
-import com.inhabas.api.auth.domain.socialAccount.type.Provider;
+import com.inhabas.api.auth.domain.oauth2.OAuth2Provider;
 import com.inhabas.api.auth.domain.socialAccount.type.UID;
 import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -10,7 +10,7 @@ import java.util.Optional;
 public interface SocialAccountRepository extends JpaRepository<SocialAccount, Integer> {
 
     @EntityGraph(attributePaths = "memberSocialAccount")
-    Optional<SocialAccount> findWithMemberIdByUidAndProvider(UID uid, Provider provider);
+    Optional<SocialAccount> findWithMemberIdByUidAndProvider(UID uid, OAuth2Provider provider);
 
-    Optional<SocialAccount> findByUidAndProvider(UID uid, Provider provider);
+    Optional<SocialAccount> findByUidAndProvider(UID uid, OAuth2Provider provider);
 }
