--- auth-module/src/main/java/com/inhabas/api/auth/domain/socialAccount/type/Provider.java
@@ -1,13 +1,0 @@
-package com.inhabas.api.auth.domain.socialAccount.type;
-
-public enum Provider {
-    GOOGLE("google"),
-    NAVER("naver"),
-    KAKAO("kakao");
-
-    private final String name;
-
-    Provider(String name) {
-        this.name = name;
-    }
-}
