+++ auth-module/src/test/java/com/inhabas/api/auth/domain/oauth2/OAuth2ProviderTest.java
@@ -1,6 +1,5 @@
-package com.inhabas.api.auth.domain.wrapper;
+package com.inhabas.api.auth.domain.oauth2;
 
-import com.inhabas.api.auth.domain.socialAccount.type.Provider;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -8,13 +7,13 @@ import java.util.Arrays;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class ProviderTest {
+public class OAuth2ProviderTest {
 
     @DisplayName("Provider 길이는 30자 초과하면 안된다.")
     @Test
     public void ProviderLengthMustNotBeGreaterThan30() {
         assertThat(
-            Arrays.stream(Provider.values())
+            Arrays.stream(OAuth2Provider.values())
                     .filter(provider -> provider.toString().length() > 30)
                     .findAny()).isEmpty();
     }
@@ -23,7 +22,7 @@ public class ProviderTest {
     @Test
     public void ProviderValueMustNotBeBlank() {
         assertThat(
-                Arrays.stream(Provider.values())
+                Arrays.stream(OAuth2Provider.values())
                         .filter(provider -> provider.toString().isBlank())
                         .findAny()).isEmpty();
     }
